classes are placed in separate files, but buttons of move do not work
