+++ Server/Point.js
@@ -1,0 +1,8 @@
+//класс точка с координатами в прямоугольной декартовой системе на плоскости\
+class Point {
+    constructor(x, y) {
+        this.x = x;
+        this.y = y;
+    }
+}
+module.exports = Point;
